Заменяет моковые данные на данные с сервера
